eval: rexp3 should operate with 64bit numbers
reloc_value returns 64bit numbers but we strip it down
to 'int' which causes problems if the former value is
big enough to overflow 'int'. Fix it.

[ BR3104312 ]

Reported-by: Christian Masloch

+++ eval.c
@@ -406,7 +406,7 @@ static expr *rexp3(int critical)
                        TOKEN_GE ? ">=" : ">"));
                 v = 0;          /* must set it to _something_ */
             } else {
-                int vv = reloc_value(e);
+                int64_t vv = reloc_value(e);
                 if (vv == 0)
                     v = (j == TOKEN_LE || j == TOKEN_GE);
                 else if (vv > 0)
